INF: Added source file for lecture INF.5.01.L01
The added file contains the code (mostly from PHP documentation) that
is presented in the "Laboratorio 1 PHP Base" on the module "Gestione
Dati" in INF.5.

The same file will be linked in the lecture on the website, therefore,
the file path should not change without unless needed (in which case
the referring should change as well).