INF: Added file for "big" database.
The file db_school_big.sql defines and populates a database that has
the same structure of the "School Database" used in the study material.
This file has a larger and more diversified set of tuples so that some
queries are better understood when applied to such a "bigger" (still
pretty small, though) database.