INF: Added DBScuola_004.
The added database DBScuola_004 contains 6 tables with various
connections. The addition w.r.t. the DBScuola_003 database is the table
`mark` that contains mark for students in a given subject by a given
teacher. The table is particularly useful to explain concepts related
to aggregate functions (GROUP BY of SQL).